requirements.txt
pandas>=2.2
openpyxl>=3.1
psycopg2-binary>=2.9
SQLAlchemy>=2.0
python-dotenv>=1.0
loguru>=0.7
